Using api classes in module_utils package.
